build table layout finished
