+++ frontend/src/components/lastRunTable/Error.styles.js
@@ -1,0 +1,6 @@
+﻿import styled from 'styled-components';
+
+export const ErrorMsg = styled.td`
+    background: ${props =>
+        props.build.status === 'FAIL' && 'var(--arabia-red) !important'};
+`;
