+++ frontend/src/components/lastRunTable/Status.jsx
@@ -1,11 +1,6 @@
 // eslint-disable-next-line
 import React from 'react';
 import { pickIcon } from '../TestIcon';
-import styled from 'styled-components';
-
-const StyledRow = styled.td`
-    text-align: center !important;
-`;
 
 const Status = ({ build, selectedBuild }) => {
     const testStatus =
@@ -14,7 +9,7 @@ const Status = ({ build, selectedBuild }) => {
             : 'EMPTY';
 
     const testStatusIcon = pickIcon(testStatus);
-    return <StyledRow>{testStatusIcon}</StyledRow>;
+    return <td>{testStatusIcon}</td>;
 };
 
 export default Status;
